auth_service/token_utils.py
from datetime import datetime, timedelta
from jose import jwt

SECRET_KEY = "supersecretkey"  # Use a secure, unique key in production
ALGORITHM = "HS256"

def generate_token(data: dict, expires_delta: timedelta) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + expires_delta
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def verify_token(token: str) -> dict:
    from jose import JWTError
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
