auth_service/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
bcrypt==4.0.1
beanie==2.0.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
elastic-transport==9.1.0
elasticsearch==9.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lazy-model==0.3.0
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
